refactor navigation package to remove v2 suffix and simplify imports
