+++ commonApp/src/commonMain/kotlin/com/stslex/wizard/bottom_bar/MainBottomAppBar.kt
@@ -9,7 +9,7 @@ import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.stslex.wizard.core.navigation.v2.Config
+import com.stslex.wizard.core.navigation.Config
 
 @Composable
 internal fun MainBottomAppBar(
