--- core/navigation/src/commonMain/kotlin/com/stslex/wizard/core/navigation/NavExt.kt
@@ -1,15 +1,0 @@
-package com.stslex.wizard.core.navigation
-
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
-import androidx.navigation.toRoute
-
-inline fun <reified S : Screen> NavGraphBuilder.navScreen(
-    noinline content: @Composable (S) -> Unit
-) {
-    composable<S> { backStackEntry ->
-        content(backStackEntry.toRoute())
-    }
-}
-
