+++ core/navigation/src/commonMain/kotlin/com/stslex/wizard/core/navigation/Target.kt
@@ -1,0 +1,3 @@
+package com.stslex.wizard.core.navigation
+
+interface Target
