--- core/navigation/src/commonMain/kotlin/com/stslex/wizard/core/navigation/navigator/Navigator.kt
@@ -1,13 +1,0 @@
-package com.stslex.wizard.core.navigation.navigator
-
-import com.stslex.wizard.core.navigation.Screen
-
-interface Navigator {
-
-    fun navTo(
-        screen: Screen,
-        options: NavigatorOptions = NavigatorOptions()
-    )
-
-    fun popBack()
-}
