--- core/navigation/src/commonMain/kotlin/com/stslex/wizard/core/navigation/navigator/NavigatorOptions.kt
@@ -1,5 +1,0 @@
-package com.stslex.wizard.core.navigation.navigator
-
-data class NavigatorOptions(
-    val isSingleTop: Boolean = false,
-)
