--- core/navigation/src/commonMain/kotlin/com/stslex/wizard/core/navigation/v2/Target.kt
@@ -1,3 +1,0 @@
-package com.stslex.wizard.core.navigation.v2
-
-interface Target
