+++ feature/film_feed/src/commonMain/kotlin/com/stslex/wizard/feature/film_feed/navigation/FilmFeedComponent.kt
@@ -1,8 +1,8 @@
 package com.stslex.wizard.feature.film_feed.navigation
 
 import com.arkivanov.decompose.ComponentContext
-import com.stslex.wizard.core.navigation.v2.Component
-import com.stslex.wizard.core.navigation.v2.Config
+import com.stslex.wizard.core.navigation.Component
+import com.stslex.wizard.core.navigation.Config
 
 interface FilmFeedComponent : Component {
