+++ feature/film_feed/src/commonMain/kotlin/com/stslex/wizard/feature/film_feed/navigation/FilmFeedComponentImpl.kt
@@ -1,7 +1,7 @@
 package com.stslex.wizard.feature.film_feed.navigation
 
 import com.arkivanov.decompose.ComponentContext
-import com.stslex.wizard.core.navigation.v2.Config
+import com.stslex.wizard.core.navigation.Config
 
 internal class FeedFilmComponentImpl(
     componentContext: ComponentContext,
