+++ feature/follower/src/commonMain/kotlin/com/stslex/wizard/feature/follower/ui/store/FollowerStore.kt
@@ -1,7 +1,7 @@
 package com.stslex.wizard.feature.follower.ui.store
 
 import androidx.compose.runtime.Stable
-import com.stslex.wizard.core.navigation.v2.Config
+import com.stslex.wizard.core.navigation.Config
 import com.stslex.wizard.core.ui.kit.base.paging.PagingConfig
 import com.stslex.wizard.core.ui.kit.base.paging.PagingUiState
 import com.stslex.wizard.core.ui.mvi.CommonEvents
