+++ feature/match/src/commonMain/kotlin/com/stslex/wizard/feature/match/navigation/MatchComponent.kt
@@ -1,8 +1,8 @@
 package com.stslex.wizard.feature.match.navigation
 
 import com.arkivanov.decompose.ComponentContext
-import com.stslex.wizard.core.navigation.v2.Component
-import com.stslex.wizard.core.navigation.v2.Config
+import com.stslex.wizard.core.navigation.Component
+import com.stslex.wizard.core.navigation.Config
 
 interface MatchComponent : Component {
