+++ feature/profile/src/commonMain/kotlin/com/stslex/wizard/feature/profile/navigation/ProfileComponent.kt
@@ -1,9 +1,9 @@
 package com.stslex.wizard.feature.profile.navigation
 
 import com.arkivanov.decompose.ComponentContext
-import com.stslex.wizard.core.navigation.v2.Component
-import com.stslex.wizard.core.navigation.v2.Config
-import com.stslex.wizard.core.navigation.v2.Config.BottomBar.Profile.Type
+import com.stslex.wizard.core.navigation.Component
+import com.stslex.wizard.core.navigation.Config
+import com.stslex.wizard.core.navigation.Config.BottomBar.Profile.Type
 import com.stslex.wizard.core.ui.mvi.v2.Handler
 import com.stslex.wizard.feature.profile.ui.store.ProfileHandlerStore
 import com.stslex.wizard.feature.profile.ui.store.ProfileStore.Action.Navigation
