+++ feature/profile/src/commonMain/kotlin/com/stslex/wizard/feature/profile/navigation/ProfileComponentImpl.kt
@@ -1,10 +1,10 @@
 package com.stslex.wizard.feature.profile.navigation
 
 import com.arkivanov.decompose.ComponentContext
-import com.stslex.wizard.core.navigation.v2.Config
-import com.stslex.wizard.core.navigation.v2.Config.BottomBar.Profile.Type
-import com.stslex.wizard.core.navigation.v2.Config.Follower.FollowerType.FOLLOWER
-import com.stslex.wizard.core.navigation.v2.Config.Follower.FollowerType.FOLLOWING
+import com.stslex.wizard.core.navigation.Config
+import com.stslex.wizard.core.navigation.Config.BottomBar.Profile.Type
+import com.stslex.wizard.core.navigation.Config.Follower.FollowerType.FOLLOWER
+import com.stslex.wizard.core.navigation.Config.Follower.FollowerType.FOLLOWING
 import com.stslex.wizard.feature.profile.ui.store.ProfileHandlerStore
 import com.stslex.wizard.feature.profile.ui.store.ProfileStore.Action.Navigation
