+++ feature/settings/src/commonMain/kotlin/com/stslex/wizard/feature/settings/navigation/SettingsComponentImpl.kt
@@ -1,7 +1,7 @@
 package com.stslex.wizard.feature.settings.navigation
 
 import com.arkivanov.decompose.ComponentContext
-import com.stslex.wizard.core.navigation.v2.Config
+import com.stslex.wizard.core.navigation.Config
 
 internal class SettingsComponentImpl(
     context: ComponentContext,
